ENH/WIP: Index[bool]

+++ pandas/tests/series/indexing/test_setitem.py
@@ -697,7 +697,7 @@ class SetitemCastingEquivalents:
         mask[key] = True
 
         res = Index(obj).where(~mask, val)
-        tm.assert_index_equal(res, Index(expected))
+        tm.assert_index_equal(res, Index(expected, dtype=expected.dtype))
 
     def test_index_putmask(self, obj, key, expected, val):
         if Index(obj).dtype != obj.dtype:
@@ -708,7 +708,7 @@ class SetitemCastingEquivalents:
         mask[key] = True
 
         res = Index(obj).putmask(mask, val)
-        tm.assert_index_equal(res, Index(expected))
+        tm.assert_index_equal(res, Index(expected, dtype=expected.dtype))
 
 
 @pytest.mark.parametrize(
